Attempting to compile project, fails
numerous

package edu.berkeley.androidwave.waverecipe.WaveRecipeAuthorization does not exist

errors

It appears that aidl is not generating classes for
files in the waverecipe package.  Need to trigger
this somehow.  May need to reorganize packages in
the main project so that all aidls for the clients
are in the same package.